pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "danish-rl"
version = "0.1.0"
description = "Danish RL training pipeline with judge-based rewards and offline prompt translation."
requires-python = ">=3.10,<3.14"
dependencies = [
  "torch>=2.5.0",
  "transformers>=4.52.0",
  "accelerate>=1.2.0",
  "datasets>=3.2.0",
  "pyyaml>=6.0.1",
  "python-dotenv>=1.0.1",
  "wandb>=0.19.0",
  "sentencepiece>=0.2.0",
  "protobuf>=4.25.0",
  "safetensors>=0.4.5",
  "nvidia-cuda-runtime-cu12>=12.4; platform_system == 'Linux'",
  "nvidia-cuda-nvrtc-cu12>=12.4; platform_system == 'Linux'",
  "nvidia-cublas-cu12>=12.4; platform_system == 'Linux'",
  "nvidia-cudnn-cu12>=9.1; platform_system == 'Linux'",
  "nvidia-cusolver-cu12>=11.6; platform_system == 'Linux'",
  "nvidia-cusparse-cu12>=12.3; platform_system == 'Linux'",
]

[project.scripts]
train-danish-rl = "train:main"
translate-prompts = "translate:main"

[tool.setuptools]
py-modules = [
  "buffer",
  "config",
  "data",
  "loss",
  "prompt",
  "prompts",
  "reward_judge",
  "runtime_env",
  "train",
  "translate",
]
